requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
idna==2.8
joblib==0.13.2
mimesis==3.3.0
numpy==1.17.1
pandas==0.25.1
psycopg2==2.7.7
psycopg2-binary==2.8.3
pycparser==2.19
pyfcm==1.4.7
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
queries==2.0.1
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tornado==6.0.3
urllib3==1.25.3
